code changes to return error if next 10 segment are not avaialble.
This change is made to support stream inerruption use case. If
segments are not avaialble because of UDP stream interruption,
read packet is stuck in demux infinitely until segments become
avaialble.

This chnage will sure that it retries 10 times to get a packet
from segments and if not avaialble, returns an error and the
packager will close because of this error.